Refactor forcing config to a submodule
