+++ tests/black_CrocoDash.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 
 
-black CrocoDash/extract_forcings CrocoDash/raw_data_access tests CrocoDash/case.py CrocoDash/forcing_configurations.py CrocoDash/logging.py
+black CrocoDash/extract_forcings CrocoDash/raw_data_access tests CrocoDash/case.py CrocoDash/forcing_configurations/ CrocoDash/logging.py
