+++ tests/forcing_configurations/test_forcing_configurations.py
@@ -1,4 +1,4 @@
-from CrocoDash.forcing_configurations import *
+from CrocoDash.forcing_configurations.base import *
 from ProConPy.config_var import cvars
 import pytest
 from types import SimpleNamespace
